Challenge 8: Roman Numerals Encoder
https://www.codewars.com/kata/51b62bf6a9c58071c600001b/train/python

Feedback:
This one took me a lil bit, had to write on a notebook to really understand the algorithm. Really don't know if it there is a more efficient way, think that those '[::-1}' that reverse the string are not that efficient, don't really know if they are O'(N). But i think that it is solved in a very "pythony" way, that's a new term that I learnt yesterday. Pretty fun exercise.